add: implemented admin authentication on cookies, if mongoose error comes up just reload the test

+++ server/config/db.js
@@ -52,8 +52,11 @@ const database = {
         //this function is taken from https://www.freecodecamp.org/news/end-point-testing/
         const collections = Object.keys(this.conn.collections);
         for (const collectionName of collections) {
-          const collection = this.conn.collections[collectionName];
-          await collection.deleteMany();
+            if(collectionName != "admins"){
+                const collection = this.conn.collections[collectionName];
+                await collection.deleteMany();
+            }
+            
         }
         
     },
